% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\name{carregar_inputs}
\alias{carregar_inputs}
\title{Carregar Dados}
\usage{
carregar_inputs(arquivo_de_inputs = "Dados.xlsx")
}
\arguments{
\item{arquivo_de_inputs}{Planilha do Excel Padronizada contendo Inputs do Modelo}
}
\description{
Esta funcao carrega dados que serao usados como Input.
}
\examples{
carregar_inputs ("Planilha_de_Inputs.xlsx")
}
\keyword{inputs}
