% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cba.R
\name{descontar_fluxo_de_caixa}
\alias{descontar_fluxo_de_caixa}
\title{Descontar Fluxo de Vaixa}
\usage{
descontar_fluxo_de_caixa(variaveis_a_descontar, ano_inicial, i, parametros,
  sufixo)
}
\arguments{
\item{variaveis_a_descontar}{Vetor com nome de variaveis que terao o fluxo de caixa descontado}

\item{ano_inicial}{Número do Ano inicial que sera descontado}

\item{i}{Taxa de Desconto.}

\item{parametros}{Dataframe de parametros que contem variaveis a descontar e}
}
\value{
fluco de caixa descontado
}
\description{
Calcula o Fluxo de Caixa Descontado Individual de um Conjunto de Variáveis Definido.
}
