% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\name{escrever_arquivo_inputs_outputs}
\alias{escrever_arquivo_inputs_outputs}
\title{escrever_arquivo_inputs_outputs}
\usage{

  escrever_arquivo_inputs_outputs(arquivo_de_inputs = "./tests/testthat/Dados.xlsx")
}
\arguments{
\item{arquivo_de_inputs}{: Arquivo de Inputs (em excel), que contenha duas abas com os inputs e outputs de cada funcao.}
}
\description{
Esta funcao é de uso interno, e foi criada para fins de desenvolvimento. A cada alteração na estrutura de variáveis de input das funcoes do modelo, esta funcao deve ser rodada novamente.
Esta funcao atualiza o arquivo de dados usado pela calculadora que contém os inputs e outputs de cada funcao.
}
