% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\name{escrever_template_dados}
\alias{escrever_template_dados}
\title{escrever_template_dados}
\usage{

  escrever_template_dados(arquivo_template_dados = "./tests/testthat/Dados.xlsx")
}
\arguments{
\item{arquivo_template_dados}{caminho para o arquivo que serve como template para os dados de entrada da calculadora}
}
\description{
escrever_template_dados
}
