% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executar.R
\name{executarAplicativo}
\alias{executarAplicativo}
\title{Executar Aplicativo}
\usage{
executarAplicativo()
}
\description{
Esta funcao executa o app shiny que contem uma interface grafica para a simulacao de beneficios e custos em sst.
Para executar o aplicativo basta carregar a biblioteca e executar esta funcao
}
\examples{
library(oshcba)
executarAplicativo()
}
