% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\name{filtrar_data_frame_linhas_sem_na}
\alias{filtrar_data_frame_linhas_sem_na}
\title{filtrar_data_frame_linhas_sem_na}
\usage{
filtrar_data_frame_linhas_sem_na(data_frame_input)
}
\arguments{
\item{data_frame_input}{data.frame com variaveis a serem filtradas}
}
\value{
dataframe filtrado, removendo NAs e valores com length == 0
}
\description{
filtrar_data_frame_linhas_sem_na
}
