% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_dados.R
\name{gerar_list_dados_tratados}
\alias{gerar_list_dados_tratados}
\title{gerar_list_dados_tratados}
\usage{
gerar_list_dados_tratados()
}
\value{
list de dados tratados no formato adequado para a função de interface de dados.
}
\description{
Esta função obtém do ambiente global uma série de objetos disponibilizados pelas rotinas de tratamento de dados, e retorna um objeto list com estes dados organizados de modo adequado para a interface de dados.
}
