% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analise.R
\name{grafico_box_plot_por_iniciativa}
\alias{grafico_box_plot_por_iniciativa}
\title{grafico_box_plot_por_iniciativa}
\usage{
grafico_box_plot_por_iniciativa(resultados_cbr)
}
\arguments{
\item{resultados_cbr}{dataframe com resultados formatados no modelo "CBR".}
}
\value{
grafico com cbr por iniciativa
}
\description{
grafico_box_plot_por_iniciativa
}
