% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analise.R
\name{grafico_dbg_cbr_waterfall}
\alias{grafico_dbg_cbr_waterfall}
\title{grafico_dbg_cbr_waterfall}
\usage{
grafico_dbg_cbr_waterfall(resumo, iniciativa = "Iniciativa1")
}
\arguments{
\item{resumo}{list com resumo de variáveis retornado pela funcao resumo_cba_por_categorias(resultados_cbr)}

\item{iniciativa}{nome da iniciativa para realizar o gráfico.}
}
\value{
gráfico waterfall com o CBR
}
\description{
grafico_dbg_cbr_waterfall
}
