% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-analysis.R
\name{lucro}
\alias{lucro}
\title{Funcao Lucro}
\usage{
lucro(CustoFixo, CustoVariavel, Preco, Producao, Demanda)
}
\arguments{
\item{CustoFixo}{custo fixo}

\item{CustoVariavel}{custo variavel}

\item{Preco}{preco}

\item{Producao}{producao}

\item{Demanda}{demanda}
}
\value{
lucro obtido
}
\description{
Funcao Lucro
}
