% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_parametros.R
\name{obter_cenarios}
\alias{obter_cenarios}
\title{Obter Cenários}
\usage{
obter_cenarios(Inputs)
}
\arguments{
\item{Inputs}{Objeto de Inputs (lista)}
}
\value{
cenários
}
\description{
Obter Cenários
}
