% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_dados.R
\name{obter_constantes}
\alias{obter_constantes}
\title{obter_constantes}
\usage{
obter_constantes(template_dados, abas_a_ler, nomes_inputs, list_dados_tratados)
}
\arguments{
\item{abas_a_ler}{vetor com abas a ler}

\item{nomes_inputs}{vetor com os nomes dos inputs}

\item{list_dados_tratados}{list com dados tratados}

\item{arquivo_template}{caminho para o arquivo template de dados}
}
\value{
dataframe com constantes
}
\description{
obter_constantes
}
