% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_dados.R
\name{obter_historicoFAP_template}
\alias{obter_historicoFAP_template}
\title{obter_historicoFAP_template}
\usage{
obter_historicoFAP_template(template_dados, abas_a_ler, nomes_inputs,
  list_dados_tratados, cenario_as_is, iniciativas_a_simular)
}
\arguments{
\item{abas_a_ler}{vetor com abas a ler}

\item{nomes_inputs}{vetor com nomes de inputs}

\item{list_dados_tratados}{list gerada pela rotina de tratamento de dados}

\item{cenario_as_is}{caractér cenário as is}

\item{iniciativas_a_simular}{vetor de iniciativas a simular}

\item{arquivo_template}{caminho do arquivo de template a usar}
}
\value{
data.frame com dois anos de histórico do FAP para a simulação
}
\description{
obter_historicoFAP_template
}
