% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_dados.R
\name{obter_inputs_list_dados_tratados}
\alias{obter_inputs_list_dados_tratados}
\title{obter_inputs_list_dados_tratados}
\usage{
obter_inputs_list_dados_tratados(arquivo_template,
  list_dados_tratados = gerar_list_dados_tratados(),
  abas_a_ler = oshcba_options$abas_a_ler,
  nomes_inputs = oshcba_options$nomes_inputs)
}
\arguments{
\item{arquivo_template}{caminho para arquivo de dados a ser usado como template}

\item{list_dados_tratados}{list com os dados tratados}

\item{abas_a_ler}{abas a ler}

\item{nomes_inputs}{nomes dos inputs a atribuir às abas}
}
\value{
list com inputs transformados a partir dos dados tratados pelo script de tratamento de dados.
}
\description{
obter_inputs_list_dados_tratados
}
