% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_parametros.R
\name{obter_parametros}
\alias{obter_parametros}
\title{Obter Parametros}
\usage{
obter_parametros(Inputs, rep = 1000)
}
\arguments{
\item{Inputs}{Inputs Carregados com a funcao carregar_inputs()}
}
\value{
Dataframe com parametros para simulacao (incluindo parametros com distribuicao e dados projetados).
}
\description{
Obter Parametros
}
