% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs.R
\name{oshcba.adicionar_log}
\alias{oshcba.adicionar_log}
\title{oshcba.adicionar_log}
\usage{
oshcba.adicionar_log(string, vetor_log = oshcba.log_calculadora)
}
\arguments{
\item{string}{character com o texto a adicionar ao vetor de logs}

\item{vetor_log}{o vetor de log existente}
}
\description{
oshcba.adicionar_log
}
