% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs.R
\name{oshcba.iniciar_log}
\alias{oshcba.iniciar_log}
\title{oshcba.iniciar_log
Esta função inicia o vetor de logs da calculadora.}
\usage{
oshcba.iniciar_log()
}
\description{
oshcba.iniciar_log
Esta função inicia o vetor de logs da calculadora.
}
