% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs.R
\name{oshcba.obter_log}
\alias{oshcba.obter_log}
\title{obter_log}
\usage{
oshcba.obter_log(vetor_log = oshcba.log_calculadora)
}
\arguments{
\item{vetor_log}{vetor de logs da bibliotecao}
}
\value{
data.frame com os logs gerados
}
\description{
obter_log
}
