% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs.R
\name{oshcba.parar_execucao}
\alias{oshcba.parar_execucao}
\title{oshcba.parar_execucao}
\usage{
oshcba.parar_execucao(string, vetor_log = oshcba.log_calculadora,
  salvar_logs_csv = FALSE)
}
\arguments{
\item{string}{character com o texto a adicionar ao vetor de logs}

\item{vetor_log}{o vetor de log existente}
}
\description{
Esta funcao pode ser usada para parar a execução da biblioteca salvando um arquivo com os logs na pasta atual.
}
