% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analise.R
\name{resumo_cba_por_categorias}
\alias{resumo_cba_por_categorias}
\title{resumo_cba_por_categorias}
\usage{
resumo_cba_por_categorias(resultados_cbr)
}
\arguments{
\item{resultados_cbr}{data.frame com resultados por iniciativa (na coluna Cenario.y)}
}
\value{
list contendo duas tiblles com estatísticas por iniciativa e categoria de benefício.
}
\description{
resumo_cba_por_categorias
}
