% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analise.R
\name{resumo_cba_por_iniciativa}
\alias{resumo_cba_por_iniciativa}
\title{resumo_cbr_por_iniciativa}
\usage{
resumo_cba_por_iniciativa(resultados_cbr)
}
\arguments{
\item{resultados_cbr}{data.frame com resultados por iniciativa (na coluna Cenario.y)}
}
\value{
tiblle com estatísticas por iniciativa
}
\description{
resumo_cbr_por_iniciativa
}
