% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cba.R
\name{roi}
\alias{roi}
\title{Retorno Sobre Investimento}
\usage{
roi(costs, benefits)
}
\arguments{
\item{costs}{Custos Envolvidos}

\item{benefits}{Benefícios Envolvidos}
}
\value{
ROI (numérico)
}
\description{
Retorno Sobre Investimento
}
