% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simular.R
\name{simular_cba}
\alias{simular_cba}
\title{Simular CBA}
\usage{
simular_cba(ArquivoInputs = "./tests/testthat/Dados.xlsx", rep = 1000,
  tipo_input = "excel", modo = "customizado", output = "completo",
  verificar_inputs = TRUE)
}
\arguments{
\item{ArquivoInputs}{Arquivo de dados usado como Input (deve seguir um padrao especifico).}

\item{rep}{Numero de replicacoes (padrão 1000)}

\item{tipo_input}{"excel" ou "list". A opcao list pode ser usada para chamar a funcao pelo R.}

\item{modo}{Modo de simulacao "basico", "simplificado" ou "customizado".}

\item{output}{Tipo do output. ("completo" ou "simples"). O modo simples retorna apenas os resultados finais, enquanto o modo completo retorna todos as as variaveis de simulacao.}
}
\value{
um list com os resultados.
}
\description{
Simular CBA
}
