% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simular.R
\name{simular_externo}
\alias{simular_externo}
\title{simular_externo
Esta funcao "encapsula" a funcao de simulacao e a executa incluindo logs}
\usage{
simular_externo(ArquivoInputs = "./tests/testthat/Dados.xlsx",
  tipo_input = "excel", modo = "customizado", output = "completo")
}
\arguments{
\item{ArquivoInputs}{inputs de simulacao a usar}

\item{tipo_input}{tipo de input}

\item{modo}{modo de simulacao}

\item{output}{tipo de output desejado}
}
\value{
list com resultados
}
\description{
simular_externo
Esta funcao "encapsula" a funcao de simulacao e a executa incluindo logs
}
