% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analise.R
\name{tabela_soma_razao_beneficio_custo}
\alias{tabela_soma_razao_beneficio_custo}
\title{tabela_soma_razao_beneficio_custo}
\usage{
tabela_soma_razao_beneficio_custo(resultados_cbr)
}
\arguments{
\item{resultados_cbr}{dataframe com resultados formatados no modelo "CBR".}
}
\value{
dataframe com resumo do CBR por iniciativa e razao beneficio custo total
}
\description{
tabela_soma_razao_beneficio_custo
}
