% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cba.R
\name{valor_presente}
\alias{valor_presente}
\title{Valor Presente}
\usage{
valor_presente(fc, i)
}
\arguments{
\item{fc}{Fluxo de Caixa a ser descontado (vetor)}

\item{i}{Taxa de Retorno a ser utilizada (numerico)}
}
\value{
Vetor do Fluxo de Caixa Descontado
}
\description{
Valor Presente
}
\examples{
valor_presente (1:10,0.05)
}
