% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-processamento.R
\name{verificar_coerencia_parametros_aleatorios}
\alias{verificar_coerencia_parametros_aleatorios}
\title{verificar_coerencia_parametros_aleatorios}
\usage{
verificar_coerencia_parametros_aleatorios(inputs)
}
\arguments{
\item{inputs}{list de inputs lidos pelo modelo.}
}
\value{
booleano informando se há inconsistência ou não nos parâmetros aleatórios
}
\description{
verificar_coerencia_parametros_aleatorios
}
