% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simular.R
\name{calcular_funcoes}
\alias{calcular_funcoes}
\title{Calcular Funcoes}
\usage{
calcular_funcoes(parametros, inputs_funcoes, output_funcoes, funcoes)
}
\arguments{
\item{parametros}{Dataframe de parametros a serem usados para o calculo}

\item{inputs_funcoes}{Dataframe de inputs por Funcao}

\item{output_funcoes}{Dataframe de Outputs por Funcao}

\item{funcoes}{Vetor com o nome das Funcoes a serem calculadas}
}
\value{
Dataframe de resultados com as variaveis calculadas
}
\description{
Calcular Funcoes
}
