% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\name{carregar_inputs}
\alias{carregar_inputs}
\title{carregar_inputs}
\usage{
carregar_inputs(arquivo_de_inputs = "./tests/testthat/Dados.xlsx",
  abas_a_ler = oshcba_options$abas_a_ler,
  nomes_inputs = oshcba_options$nomes_inputs)
}
\arguments{
\item{arquivo_de_inputs}{arquivo excel de inputs}

\item{abas_a_ler}{vetor com abas a ler}

\item{nomes_inputs}{vetor com nomes dos inputs}
}
\value{
data.frame com os dados lidos.
}
\description{
carregar_inputs
}
