% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cba.R
\name{cbr}
\alias{cbr}
\title{Razão Benefício Custo}
\usage{
cbr(costs, benefits)
}
\arguments{
\item{costs}{Os Custos da Iniciativa em Valores Monetários (numérico).}

\item{benefits}{O Benefício em Valores Monetários (numérico).}
}
\value{
Razão Custo Benefício
}
\description{
Esta função Calcula a razão Benefício Custo.
}
\examples{
cbr(20,30)
}
\keyword{cbr}
