% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cba.R
\name{cbr}
\alias{cbr}
\title{Razao Beneficio Custo}
\usage{
cbr(costs, benefits)
}
\arguments{
\item{costs}{Os Custos da Iniciativa em Valores Monetarios (numerico).}

\item{benefits}{O Beneficio em Valores Monetarios (numerico).}
}
\value{
Razao Beneficio Custo
}
\description{
Esta funcao Calcula a razao beneficio custo
}
\examples{
cbr(20,30)
}
\keyword{cbr}
