% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cba.R
\name{cbr}
\alias{cbr}
\title{Razão Benefício Custo}
\usage{
cbr(costs, benefits)
}
\arguments{
\item{costs}{Os Custos da Iniciativa em Valores Monetários (numérico).}

\item{benefits}{O Benefício em Valores Monetários (numérico).}
}
\value{
Razão Benefício Custo
}
\description{
Esta função Calcula a razão benefício custo
}
\keyword{cbr}
