% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_parametros.R
\name{obter_cenarios}
\alias{obter_cenarios}
\title{Obter Cenarios}
\usage{
obter_cenarios(Inputs)
}
\arguments{
\item{Inputs}{Objeto de Inputs (lista)}
}
\value{
cenarios
}
\description{
Obter Cenarios
}
