% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_parametros.R
\name{obter_parametros}
\alias{obter_parametros}
\title{Obter Parâmetros}
\usage{
obter_parametros(Inputs, rep = 1000)
}
\arguments{
\item{Inputs}{Inputs Carregados com a função carregar_inputs()}
}
\value{
Dataframe com parâmetros para simulação (incluindo parâmetros com distribuição e dados projetados).
}
\description{
Obter Parâmetros
}
