% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oshcba.R
\docType{package}
\name{oshcba}
\alias{oshcba}
\alias{oshcba-package}
\title{oshcba: Um pacote para Analise de Custo Beneficio em Saude e Seguranca do Trabalho}
\description{
Este Pacote contem funcoes necessarias para uma analise de custo e beneficio probabilistica.
}
\section{Funcoes Disponiveis}{


Diversas funcoes sao incluidas em cada um dos arquivos:
Dados.R: carregar_inputs(),
cba.R:
simular.R:
output.R:
}

\author{
Pedro Nascimento de Lima
}
