% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oshcba.R
\docType{package}
\name{oshcba}
\alias{oshcba}
\alias{oshcba-package}
\title{oshcba: Um pacote para Análise de Custo Benefício em Saúde e Segurança do Trabalho.}
\description{
Este Pacote contém funções necessárias para uma análise de custo e benefício probabilística.
}
\section{Funções Disponíveis}{


Diversas funçoes são incluídas em cada um dos arquivos:
Dados.R: carregar_inputs(),
cba.R:
simular.R:
output.R:
}

\author{
Pedro Nascimento de Lima
}
