% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oshcba.R
\docType{package}
\name{oshcba}
\alias{oshcba}
\alias{oshcba-package}
\title{oshcba: Um pacote para Análise Custo Benefício em Saúde e Segurança do Trabalho}
\description{
Este Pacote contém as funções necessárias para uma análise Custo Benefício.
}
\section{Funções Disponíveis}{


Diversas funções São incluídas em cada um dos Arquivos:
Dados.R: carregar_inputs(),
cba.R:
simular.R:
output.R:
}

\author{
Pedro Nascimento de Lima
}
