% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simular.R
\name{simular_cba}
\alias{simular_cba}
\title{Simular CBA}
\usage{
simular_cba(ArquivoInputs = "./tests/testthat/Dados.xlsx",
  tipo_input = "excel", modo = "customizado", output = "completo")
}
\arguments{
\item{ArquivoInputs}{Arquivo de dados usado como Input (deve seguir um padrao especifico).}

\item{modo}{Modo de simulacao ("completo" ou "simples"). O modo simples retorna apenas os resultados finais, enquanto o modo completo retorna todos as as variaveis de simulacao.}
}
\value{
um list com os resultados.
}
\description{
Simular CBA
}
