% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simular.R
\name{simular_cba}
\alias{simular_cba}
\title{Simular CBA}
\usage{
simular_cba(ArquivoInputs = "./tests/testthat/Dados.xlsx", rep = 1000,
  tipo_input = "excel", modo = "customizado", output = "completo",
  verificar_inputs = FALSE)
}
\arguments{
\item{ArquivoInputs}{Arquivo de dados usado como Input (deve seguir um padrão específico).}

\item{rep}{Número de replicações (padrão 1000)}

\item{tipo_input}{"excel" ou "list". A opção list pode ser usada para chamar a função pelo R.}

\item{modo}{Modo de simulacao "básico", "simplificado" ou "customizado".}

\item{output}{Tipo do output. ("completo" ou "simples"). O modo simples retorna apenas os resultados finais, enquanto o modo completo retorna todas as variáveis de simulação.}
}
\value{
um list com os resultados.
}
\description{
Simular CBA
}
