% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simular.R
\name{simular_externo}
\alias{simular_externo}
\title{simular_externo}
\usage{
simular_externo(ArquivoInputs = "./tests/testthat/Dados.xlsx",
  tipo_input = "excel", modo = "customizado", output = "completo")
}
\arguments{
\item{output}{}
}
\value{
list com resultados
}
\description{
simular_externo
}
