% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cba.R
\name{valor_presente}
\alias{valor_presente}
\title{Fluxo de Caixa Descontado}
\usage{
valor_presente(fc, t, i)
}
\arguments{
\item{fc}{Fluxo de Caixa (numero, em valores monetarios)}

\item{t}{Tempo do fluxo de caixa.}

\item{i}{Taxa de Juros.}
}
\value{
fcd Fluxo de Caixa Descontado
}
\description{
Fluxo de Caixa Descontado
}
